Renaming rows - the old (original) "rows" to "LegacyRow" - the new temporary "row2" to new "row"
