+++ lib/Service/RowService.php
@@ -4,10 +4,10 @@ namespace OCA\Tables\Service;
 
 use OCA\Tables\Db\Column;
 use OCA\Tables\Db\ColumnMapper;
+use OCA\Tables\Db\LegacyRow;
 use OCA\Tables\Db\Row;
-use OCA\Tables\Db\Row2;
-use OCA\Tables\Db\Row2Mapper;
 use OCA\Tables\Db\RowMapper;
+use OCA\Tables\Db\LegacyRowMapper;
 use OCA\Tables\Db\Table;
 use OCA\Tables\Db\TableMapper;
 use OCA\Tables\Db\View;
@@ -29,15 +29,15 @@ use Psr\Log\LoggerInterface;
  * @psalm-import-type TablesRow from ResponseDefinitions
  */
 class RowService extends SuperService {
-	private RowMapper $mapper;
+	private LegacyRowMapper $mapper;
 	private ColumnMapper $columnMapper;
 	private ViewMapper $viewMapper;
 	private TableMapper $tableMapper;
-	private Row2Mapper $row2Mapper;
+	private RowMapper $row2Mapper;
 	private array $tmpRows = []; // holds already loaded rows as a small cache
 
 	public function __construct(PermissionsService $permissionsService, LoggerInterface $logger, ?string $userId,
-		RowMapper $mapper, ColumnMapper $columnMapper, ViewMapper $viewMapper, TableMapper $tableMapper, Row2Mapper $row2Mapper) {
+								LegacyRowMapper    $mapper, ColumnMapper $columnMapper, ViewMapper $viewMapper, TableMapper $tableMapper, RowMapper $row2Mapper) {
 		parent::__construct($logger, $userId, $permissionsService);
 		$this->mapper = $mapper;
 		$this->columnMapper = $columnMapper;
@@ -47,11 +47,11 @@ class RowService extends SuperService {
 	}
 
 	/**
-	 * @param Row2[] $rows
+	 * @param Row[] $rows
 	 * @psalm-return TablesRow[]
 	 */
 	public function formatRows(array $rows): array {
-		return array_map(fn (Row2 $row) => $row->jsonSerialize(), $rows);
+		return array_map(fn (Row $row) => $row->jsonSerialize(), $rows);
 	}
 
 	/**
@@ -59,7 +59,7 @@ class RowService extends SuperService {
 	 * @param string $userId
 	 * @param ?int $limit
 	 * @param ?int $offset
-	 * @return Row2[]
+	 * @return Row[]
 	 * @throws InternalError
 	 * @throws PermissionError
 	 */
@@ -83,7 +83,7 @@ class RowService extends SuperService {
 	 * @param string $userId
 	 * @param int|null $limit
 	 * @param int|null $offset
-	 * @return Row2[]
+	 * @return Row[]
 	 * @throws DoesNotExistException
 	 * @throws InternalError
 	 * @throws MultipleObjectsReturnedException
@@ -110,12 +110,12 @@ class RowService extends SuperService {
 
 	/**
 	 * @param int $id
-	 * @return Row2
+	 * @return Row
 	 * @throws InternalError
 	 * @throws NotFoundError
 	 * @throws PermissionError
 	 */
-	public function find(int $id): Row2 {
+	public function find(int $id): Row {
 		try {
 			$columns = $this->columnMapper->findAllByTable($id);
 		} catch (Exception $e) {
@@ -145,14 +145,14 @@ class RowService extends SuperService {
 	 * @param int|null $tableId
 	 * @param int|null $viewId
 	 * @param list<array{columnId: int, value: mixed}> $data
-	 * @return Row2
+	 * @return Row
 	 *
 	 * @throws NotFoundError
 	 * @throws PermissionError
 	 * @throws Exception
 	 * @throws InternalError
 	 */
-	public function create(?int $tableId, ?int $viewId, array $data): Row2 {
+	public function create(?int $tableId, ?int $viewId, array $data): Row {
 		if ($this->userId === null || $this->userId === '') {
 			$e = new \Exception('No user id in context, but needed.');
 			$this->logger->error($e->getMessage(), ['exception' => $e]);
@@ -200,7 +200,7 @@ class RowService extends SuperService {
 		$data = $this->cleanupData($data, $columns, $tableId, $viewId);
 
 		// perf
-		$row2 = new Row2();
+		$row2 = new Row();
 		$row2->setTableId($tableId);
 		$row2->setData($data);
 		try {
@@ -296,7 +296,7 @@ class RowService extends SuperService {
 	 * @throws NotFoundError
 	 * @throws InternalError
 	 */
-	private function getRowById(int $rowId): Row2 {
+	private function getRowById(int $rowId): Row {
 		if (isset($this->tmpRows[$rowId])) {
 			return $this->tmpRows[$rowId];
 		}
@@ -327,7 +327,7 @@ class RowService extends SuperService {
 	 * @param int|null $viewId
 	 * @param list<array{columnId: string|int, value: mixed}> $data
 	 * @param string $userId
-	 * @return Row2
+	 * @return Row
 	 *
 	 * @throws InternalError
 	 * @throws NotFoundError
@@ -338,7 +338,7 @@ class RowService extends SuperService {
 		?int $viewId,
 		array $data,
 		string $userId
-	): Row2 {
+	): Row {
 		try {
 			$item = $this->getRowById($id);
 		} catch (InternalError $e) {
@@ -418,14 +418,14 @@ class RowService extends SuperService {
 	 * @param int $id
 	 * @param int|null $viewId
 	 * @param string $userId
-	 * @return Row2
+	 * @return Row
 	 *
 	 * @throws InternalError
 	 * @throws NotFoundError
 	 * @throws PermissionError
 	 * @noinspection DuplicatedCode
 	 */
-	public function delete(int $id, ?int $viewId, string $userId): Row2 {
+	public function delete(int $id, ?int $viewId, string $userId): Row {
 		try {
 			$item = $this->getRowById($id);
 		} catch (InternalError $e) {
